update: separate and rename adapter class
