+++ app/src/main/res/layout/fragment_randomgame.xml
@@ -21,7 +21,7 @@
         />
 
     <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/my_recycler_view"
+        android:id="@+id/rg_recycler_view"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_weight="1"
